Extracted data fetcher invocation logic out of SchemaProvider
